fix: move common func to utils
