+++ docs/en/latest/control-api.md
@@ -476,8 +476,10 @@ Dumps the metadata with the specified `plugin_name`:
 }
 ```
 
-### POST /v1/plugins_reload
+### PUT /v1/plugins/reload
 
 Introduced in [v3.9.0](https://github.com/apache/apisix/releases/tag/3.9.0)
 
 Triggers a hot reload of the plugins.
+
+```shell
